Refactor: extract shared IBOM into a component.
The same structure is shared between `[username]/[projectName]/IBOM` and
`[username]/[porjectName]/[multiProjectName]/IBOM`.
