+++ frontend/src/components/Board/BoardExtrasMenu.js
@@ -1,13 +1,15 @@
 import React from 'react'
 import Link from 'next/link'
+import { useRouter } from 'next/router'
 import { Button, Header } from 'semantic-ui-react'
 
 import { bool, string } from 'prop-types'
 import IBomIcon from './IBomIcon'
 import TracespaceIcon from './TracespaceIcon'
 
-const BoardExtraMenus = ({ zipUrl, hasInteractiveBom, projectFullname }) => {
-  const ibomUrl = `/${projectFullname}/IBOM/`
+const BoardExtraMenus = ({ zipUrl, hasInteractiveBom }) => {
+  const { asPath } = useRouter()
+  const ibomUrl = `${asPath}/IBOM/`
 
   return (
     <div
@@ -43,7 +45,6 @@ const BoardExtraMenus = ({ zipUrl, hasInteractiveBom, projectFullname }) => {
 BoardExtraMenus.propTypes = {
   zipUrl: string.isRequired,
   hasInteractiveBom: bool.isRequired,
-  projectFullname: string.isRequired,
 }
 
 export default BoardExtraMenus
