+++ frontend/src/pages/[username]/[projectName]/[multiProjectName]/index.js
@@ -77,7 +77,6 @@ export const getServerSideProps = async ({ params, query, req }) => {
         readmeExists: readmeFile !== '',
         kitspaceYAMLExists,
         finishedProcessing,
-        multiProjectName,
       },
     }
   }
